requirements.txt
Django==1.6
PIL==1.1.7
boto==2.22.1
dj-database-url==0.2.2
dj-static==0.0.5
django-absolute==0.3
django-annoying==0.7.9
django-cleanup==0.1.8
django-storages==1.1.8
gunicorn==18.0
psycopg2==2.4.6
static==0.4
wsgiref==0.1.2
